requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
dataclasses==0.6
environs==9.5.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
marshmallow==3.19.0
multidict==6.0.4
packaging==23.1
postgres==4.0
psycopg2-binary==2.9.6
psycopg2-pool==1.1
pycodestyle==2.10.0
pydantic==1.10.9
python-dotenv==1.0.0
typing_extensions==4.7.0
urllib3==2.0.3
yarl==1.9.2
